Move keyboardShortcuts and theme up one level
