+++ shared/src/commandPalette/CommandList.tsx
@@ -11,7 +11,7 @@ import TooltipPopoverWrapper from 'reactstrap/lib/TooltipPopoverWrapper'
 import { Subscription } from 'rxjs'
 import stringScore from 'string-score'
 import { Key } from 'ts-key-enum'
-import { KeyboardShortcut } from '../keyboardShortcuts/keyboardShortcuts'
+import { KeyboardShortcut } from '../keyboardShortcuts'
 import { ActionItem, ActionItemAction } from '../actions/ActionItem'
 import { ContributableMenu, Contributions, Evaluated } from '../api/protocol'
 import { HighlightedMatches } from '../components/HighlightedMatches'
