+++ shared/src/components/FileMatchChildren.tsx
@@ -2,7 +2,7 @@ import H from 'history'
 import { flatMap } from 'lodash'
 import * as React from 'react'
 import { Observable } from 'rxjs'
-import { ThemeProps } from '../theme/theme'
+import { ThemeProps } from '../theme'
 import { isSettingsValid, SettingsCascadeProps } from '../settings/settings'
 import { SymbolIcon } from '../symbols/SymbolIcon'
 import { toPositionOrRangeHash } from '../util/url'
