+++ web/src/components/SearchResultMatch.tsx
@@ -13,7 +13,7 @@ import { highlightNode } from '../../../shared/src/util/dom'
 import { renderMarkdown } from '../discussions/backend'
 import { highlightCode } from '../search/backend'
 import { HighlightRange } from './SearchResult'
-import { ThemeProps } from '../../../shared/src/theme/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 
 interface SearchResultMatchProps extends ThemeProps {
     item: GQL.ISearchResultMatch
