+++ web/src/enterprise/main.tsx
@@ -10,7 +10,7 @@ import '../sentry'
 import React from 'react'
 import { render } from 'react-dom'
 import '../enterprise.scss'
-import { KEYBOARD_SHORTCUTS } from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
+import { KEYBOARD_SHORTCUTS } from '../../../shared/src/keyboardShortcuts'
 import { SourcegraphWebApp } from '../SourcegraphWebApp'
 import { enterpriseExploreSections } from './explore/exploreSections'
 import { enterpriseExtensionAreaHeaderNavItems } from './extensions/extension/extensionAreaHeaderNavItems'
