+++ web/src/explore/ExploreArea.tsx
@@ -6,7 +6,7 @@ import { SettingsCascadeOrError } from '../../../shared/src/settings/settings'
 import { eventLogger } from '../tracking/eventLogger'
 import { ComponentDescriptor } from '../util/contributions'
 import { PatternTypeProps } from '../search'
-import { ThemeProps } from '../../../shared/src/theme/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 
 /**
  * Properties passed to all section components in the explore area.
