+++ web/src/main.tsx
@@ -15,7 +15,7 @@ import { extensionAreaRoutes } from './extensions/extension/routes'
 import { extensionsAreaHeaderActionButtons } from './extensions/extensionsAreaHeaderActionButtons'
 import { extensionsAreaRoutes } from './extensions/routes'
 import './main.scss'
-import { KEYBOARD_SHORTCUTS } from '../../shared/src/keyboardShortcuts/keyboardShortcuts'
+import { KEYBOARD_SHORTCUTS } from '../../shared/src/keyboardShortcuts'
 import { orgAreaHeaderNavItems } from './org/area/navitems'
 import { orgAreaRoutes } from './org/area/routes'
 import { repoHeaderActionButtons } from './repo/repoHeaderActionButtons'
