+++ web/src/marketing/SurveyPage.tsx
@@ -12,7 +12,7 @@ import { eventLogger } from '../tracking/eventLogger'
 import { submitSurvey } from './backend'
 import { SurveyCTA } from './SurveyToast'
 import { Subscription } from 'rxjs'
-import { ThemeProps } from '../../../shared/src/theme/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 interface SurveyFormProps {
     location: H.Location
     history: H.History
