+++ web/src/namespaces/NamespaceArea.tsx
@@ -2,7 +2,7 @@ import { ExtensionsControllerProps } from '../../../shared/src/extensions/contro
 import * as GQL from '../../../shared/src/graphql/schema'
 import { RouteDescriptor } from '../util/contributions'
 import { PatternTypeProps } from '../search'
-import { ThemeProps } from '../../../shared/src/theme/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 
 /**
  * Properties passed to all page components in the namespace area.
