+++ web/src/nav/GlobalNavbar.tsx
@@ -8,13 +8,13 @@ import * as GQL from '../../../shared/src/graphql/schema'
 import { PlatformContextProps } from '../../../shared/src/platform/context'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { authRequired } from '../auth'
-import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
+import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts'
 import { parseSearchURLQuery, PatternTypeProps } from '../search'
 import { SearchNavbarItem } from '../search/input/SearchNavbarItem'
 import { EventLoggerProps } from '../tracking/eventLogger'
 import { showDotComMarketing } from '../util/features'
 import { NavLinks } from './NavLinks'
-import { ThemeProps, ThemePreferenceProps } from '../../../shared/src/theme/theme'
+import { ThemeProps, ThemePreferenceProps } from '../../../shared/src/theme'
 
 interface Props
     extends SettingsCascadeProps,
