+++ web/src/repo/blob/LineDecorationAttachment.tsx
@@ -4,7 +4,7 @@ import { DecorationAttachmentRenderOptions } from 'sourcegraph'
 import { decorationAttachmentStyleForTheme } from '../../../../shared/src/api/client/services/decoration'
 import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 import { AbsoluteRepoFile } from '../../../../shared/src/util/url'
-import { ThemeProps } from '../../../../shared/src/theme/theme'
+import { ThemeProps } from '../../../../shared/src/theme'
 
 interface LineDecorationAttachmentProps extends AbsoluteRepoFile, ThemeProps {
     line: number
