+++ web/src/search/input/SearchPage.tsx
@@ -8,7 +8,7 @@ import { Form } from '../../components/Form'
 import { PageTitle } from '../../components/PageTitle'
 import { Notices } from '../../global/Notices'
 import { QuickLink, Settings } from '../../schema/settings.schema'
-import { ThemePreferenceProps, ThemeProps } from '../../../../shared/src/theme/theme'
+import { ThemePreferenceProps, ThemeProps } from '../../../../shared/src/theme'
 import { eventLogger } from '../../tracking/eventLogger'
 import { limitString } from '../../util'
 import { submitSearch } from '../helpers'
